rework the kotlin-rest example onto suspend stormify and realign the react frontend to the new wire contract
- services are now suspend over a pooled stormify.suspending, reads borrow a connection and atomic writes take a transaction, the pool closes on shutdown and /api/health reports its counters
- the create and update dtos fold into one upsert request per entity behind a shared ref and a pk helper, pagedresponse size becomes pagesize, and the mappers move next to the dtos
- validation now covers reorder level, expected date parsing and shipment fields, and a requireStatus helper replaces the hand rolled state guards
- a commonTest service suite exercises document numbers, validation, missing deletes, insufficient stock and the one shipment per order rule, alongside a version catalog and a windows mkdir arity fix
- the react frontend absorbs the pagesize rename, the unified ref label, the pruned statuses and integer cents money handling

+++ kotlin-rest/src/commonMain/kotlin/com/example/kotlinrest/entity/Product.kt
@@ -14,7 +14,7 @@ class Product(
     var category: Category? by db(null)
     @DbField(name = "supplier_id")
     var supplier: Supplier? by db(null)
-    var unitPrice: Double by db(0.0)
+    var unitPrice: Long by db(0L)
     var reorderLevel: Int by db(0)
     var active: Boolean by db(true)
 }
