+++ kotlin-rest/src/commonMain/kotlin/com/example/kotlinrest/entity/PurchaseOrderItem.kt
@@ -12,6 +12,6 @@ class PurchaseOrderItem(
     @DbField(name = "product_id")
     var product: Product? by db(null)
     var quantity: Int by db(0)
-    var unitCost: Double by db(0.0)
-    var lineTotal: Double by db(0.0)
+    var unitCost: Long by db(0L)
+    var lineTotal: Long by db(0L)
 }
